refactore taskRunStatus and runStatus
pipelinerun_test.go has these three tests:

* TestUpdateTaskRunsState
* TestUpdateRunsState
* TestUpdateTaskRunStateWithConditionChecks

These tests are validating func (state PipelineRunState) GetTaskRunsStatus()
which is part of the pipelinerunstate.go. Moving these tests to
pipelinerunstate_test.go after replacing task/run/pipeline/pipelinerun
initializations with YAMLParser.
